Try to do bus borders correctly
In the previous commit, I forgot about the issue of setting route
colors, since we can't change those through Mapbox for non-SDF icons,
which are required for the multicolor (i.e. outlined) icons we want.

It looks like it would ultimately be a huge pain, requiring both a
separate drawable resource for each route and a list of all route IDs
(or maybe names). And if it turned out that our list was incomplete,
then vehicles for that route would be invisible on the map until we
pushed out an update to fix it. I think this might be too fragile.